Update Counter.sol
Got rid of the constructor and assigned the state variable count(uint) on line 5 to 0(which was the purpose of the constructor). ALSO, once we added public in front of our count variable on line 5, Solidity automatically makes count a getter function(which gets the count of our variable(count)), so we don't need the getCount() function anymore. So cool!